+++ packages/koot/i18n/onServerRender/index.js
@@ -1,19 +1,17 @@
 import {
-    actionInit,
+    actionInit
     // actionLocales
-} from '../redux'
+} from '../redux';
 
-export default ({
-    store
-}) => {
+export default ({ store }) => {
     // console.log(1)
 
     if (typeof store !== 'object' && typeof store.getState !== 'function')
-        return
+        return;
 
-    store.dispatch(actionInit(store.getState()))
-    // if (JSON.parse(process.env.KOOT_I18N_TYPE) === 'redux')
+    store.dispatch(actionInit(store.getState()));
+    // if (JSON.parse(process.env.KOOT_I18N_TYPE) === 'store')
     //     store.dispatch(actionLocales(store.getState()))
 
     // console.log(2)
-}
+};