+++ packages/koot/i18n/server/get-ssr-state.js
@@ -5,22 +5,26 @@
  * @returns {Object}
  */
 const getSSRState = () => {
-    if (!__SERVER__) return {}
+    if (!__SERVER__) return {};
     if (__DEV__)
         return {
             localeId: global.__KOOT_SSR__.LocaleId,
-            locales: JSON.parse(process.env.KOOT_I18N_TYPE) === 'redux'
-                ? (global.__KOOT_SSR__.locales[global.__KOOT_SSR__.LocaleId] || {})
-                : {}
-        }
+            locales:
+                JSON.parse(process.env.KOOT_I18N_TYPE) === 'store'
+                    ? global.__KOOT_SSR__.locales[
+                          global.__KOOT_SSR__.LocaleId
+                      ] || {}
+                    : {}
+        };
 
-    if (typeof __KOOT_SSR__ !== 'object') return {}
+    if (typeof __KOOT_SSR__ !== 'object') return {};
 
     return {
         localeId: __KOOT_SSR__.LocaleId,
-        locales: JSON.parse(process.env.KOOT_I18N_TYPE) === 'redux'
-            ? (__KOOT_SSR__.locales[__KOOT_SSR__.LocaleId] || {})
-            : {}
-    }
-}
-export default getSSRState
+        locales:
+            JSON.parse(process.env.KOOT_I18N_TYPE) === 'store'
+                ? __KOOT_SSR__.locales[__KOOT_SSR__.LocaleId] || {}
+                : {}
+    };
+};
+export default getSSRState;