+++ test/projects/standard/koot.config.js
@@ -27,15 +27,21 @@ module.exports = {
     cookiesToStore: 'all',
     sessionStore: true,
 
-    i18n: {
-        // type: 'redux', // 仅影响 client-prod 环境
-        // use: 'router',
-        locales: [
-            ['zh', './src/locales/zh.json'],
-            ['zh-tw', './src/locales/zh-tw.json'],
-            ['en', './src/locales/en.json']
-        ]
-    },
+    // i18n: {
+    //     // type: 'redux', // 仅影响 client-prod 环境
+    //     // type: 'store', // 仅影响 client-prod 环境
+    //     // use: 'router',
+    //     locales: [
+    //         ['zh', './src/locales/zh.json'],
+    //         ['zh-tw', './src/locales/zh-tw.json'],
+    //         ['en', './src/locales/en.json']
+    //     ]
+    // },
+    i18n: [
+        ['zh', './src/locales/zh.json'],
+        ['zh-tw', './src/locales/zh-tw.json'],
+        ['en', './src/locales/en.json']
+    ],
 
     pwa: true,